setup-environment-internal: choose only the image deploy dir
Switch to default bitbake deploy dir to be more aligned with other layers.

Recently meta-ti intruduces changes [1][2] that requires this.
There are some SPDX collision if we use a common deploy folder
and we need to deployes the SPDX per machine in an unshared deploy.

After several tests, I concluded that is not possible to change the DEPLOY_DIR
in meta-lmp using the ti-soc machine override given that some artefacts continue
to go to the shared directory DEPLOY_DIR="${BUILDDIR}/deploy" that we use.
So the below doesn't work:
| TI_COMMON_DEPLOY = "${TOPDIR}/deploy"
| DEPLOY_DIR:ti-soc = "${TMPDIR}/deploy"

[1] https://git.yoctoproject.org/meta-ti/commit/?id=811c87bbbd324d4f176332b9e5f7c3699310cb06
[2] https://git.yoctoproject.org/meta-ti/commit/?id=488a9238140661dce8510694a05dd04ff023855d

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>